depo_formatter/requirements.txt
customtkinter
python-docx
docxtpl
pdfplumber
requests
python-dotenv
